todoapp/requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
pytz==2019.3
PyJWT
django-cors-headers
